arch/x64: create qemu chip
