--- arch/x86_64/src/qemu/Kconfig
@@ -1,4 +1,0 @@
-#
-# For a description of the syntax of this configuration file,
-# see the file kconfig-language.txt in the NuttX tools repository.
-#
