--- boards/x86_64/intel64/qemu-intel64/configs/earlyfb/defconfig
@@ -1,63 +1,0 @@
-#
-# This file is autogenerated: PLEASE DO NOT EDIT IT.
-#
-# You can use "make menuconfig" to make any modifications to the installed .config file.
-# You can then do "make savedefconfig" to generate a new defconfig file that includes your
-# modifications.
-#
-# CONFIG_SERIAL is not set
-CONFIG_ARCH="x86_64"
-CONFIG_ARCH_BOARD="qemu-intel64"
-CONFIG_ARCH_BOARD_INTEL64_QEMU=y
-CONFIG_ARCH_CHIP="intel64"
-CONFIG_ARCH_CHIP_INTEL64_QEMU=y
-CONFIG_ARCH_INTEL64_CORE_FREQ_KHZ=2600000
-CONFIG_ARCH_SIZET_LONG=y
-CONFIG_ARCH_X86_64=y
-CONFIG_BOARD_LOOPSPERMSEC=999
-CONFIG_BOOT_RUNFROMEXTSRAM=y
-CONFIG_BUILTIN=y
-CONFIG_CONSOLE_SYSLOG=y
-CONFIG_DEBUG_FEATURES=y
-CONFIG_DEBUG_PCI=y
-CONFIG_DEBUG_PCI_ERROR=y
-CONFIG_DEBUG_PCI_INFO=y
-CONFIG_DEBUG_PCI_WARN=y
-CONFIG_DEBUG_SYMBOLS=y
-CONFIG_DEFAULT_TASK_STACKSIZE=4194304
-CONFIG_EXAMPLES_HELLO=y
-CONFIG_FS_PROCFS=y
-CONFIG_IDLETHREAD_STACKSIZE=4194304
-CONFIG_INIT_ENTRYPOINT="nsh_main"
-CONFIG_LIBM=y
-CONFIG_MULTBOOT2_FB_TERM=y
-CONFIG_NSH_ARCHINIT=y
-CONFIG_NSH_BUILTIN_APPS=y
-CONFIG_NSH_DISABLE_IFCONFIG=y
-CONFIG_NSH_DISABLE_IFUPDOWN=y
-CONFIG_NSH_READLINE=y
-CONFIG_NXFONTS=y
-CONFIG_NXFONT_MONO5X8=y
-CONFIG_PCI=y
-CONFIG_PCI_QEMU_EDU=y
-CONFIG_PCI_QEMU_TEST=y
-CONFIG_PREALLOC_CHILDSTATUS=16
-CONFIG_PRIORITY_INHERITANCE=y
-CONFIG_PTHREAD_MUTEX_TYPES=y
-CONFIG_PTHREAD_STACK_MIN=4194304
-CONFIG_RAM_SIZE=268435456
-CONFIG_SCHED_CHILD_STATUS=y
-CONFIG_SCHED_HAVE_PARENT=y
-CONFIG_SCHED_IRQMONITOR=y
-CONFIG_SCHED_TICKLESS=y
-CONFIG_SCHED_TICKLESS_ALARM=y
-CONFIG_SCHED_TICKLESS_LIMIT_MAX_SLEEP=y
-CONFIG_SCHED_WAITPID=y
-CONFIG_SIG_DEFAULT=y
-CONFIG_START_DAY=3
-CONFIG_START_MONTH=3
-CONFIG_START_YEAR=2011
-CONFIG_SYSTEM_CLE=y
-CONFIG_SYSTEM_NSH=y
-CONFIG_SYSTEM_TIME64=y
-CONFIG_USEC_PER_TICK=1
