--- boards/x86_64/intel64/qemu-intel64/configs/jumbo/defconfig
@@ -1,114 +1,0 @@
-#
-# This file is autogenerated: PLEASE DO NOT EDIT IT.
-#
-# You can use "make menuconfig" to make any modifications to the installed .config file.
-# You can then do "make savedefconfig" to generate a new defconfig file that includes your
-# modifications.
-#
-CONFIG_16550_ADDRWIDTH=16
-CONFIG_16550_PCI_UART1=y
-CONFIG_16550_PCI_UART1_CLOCK=1843200
-CONFIG_16550_PCI_UART1_DEVICE=0x0002
-CONFIG_16550_PCI_UART1_VENDOR=0x1b36
-CONFIG_16550_PCI_UART=y
-CONFIG_16550_UART0=y
-CONFIG_16550_UART0_BASE=0x3f8
-CONFIG_16550_UART0_CLOCK=1843200
-CONFIG_16550_UART0_IRQ=36
-CONFIG_16550_UART0_RXBUFSIZE=16
-CONFIG_16550_UART0_SERIAL_CONSOLE=y
-CONFIG_16550_UART0_TXBUFSIZE=16
-CONFIG_ARCH="x86_64"
-CONFIG_ARCH_BOARD="qemu-intel64"
-CONFIG_ARCH_BOARD_INTEL64_QEMU=y
-CONFIG_ARCH_CHIP="intel64"
-CONFIG_ARCH_CHIP_INTEL64_QEMU=y
-CONFIG_ARCH_INTEL64_CORE_FREQ_KHZ=2600000
-CONFIG_ARCH_INTERRUPTSTACK=8192
-CONFIG_ARCH_SIZET_LONG=y
-CONFIG_ARCH_X86_64=y
-CONFIG_ARCH_X86_64_AVX=y
-CONFIG_ARCH_X86_64_FMA=y
-CONFIG_ARCH_X86_64_SSE3=y
-CONFIG_ARCH_X86_64_SSE41=y
-CONFIG_ARCH_X86_64_SSE42=y
-CONFIG_ARCH_X86_64_SSE4A=y
-CONFIG_ARCH_X86_64_SSSE3=y
-CONFIG_BOARD_LOOPSPERMSEC=999
-CONFIG_BOOT_RUNFROMEXTSRAM=y
-CONFIG_BUILTIN=y
-CONFIG_CONSOLE_SYSLOG=y
-CONFIG_DEBUG_FEATURES=y
-CONFIG_DEBUG_PCI=y
-CONFIG_DEBUG_PCI_ERROR=y
-CONFIG_DEBUG_PCI_INFO=y
-CONFIG_DEBUG_PCI_WARN=y
-CONFIG_DEBUG_SYMBOLS=y
-CONFIG_DEFAULT_TASK_STACKSIZE=4194304
-CONFIG_DEV_SIMPLE_ADDRENV=y
-CONFIG_EXAMPLES_HELLO=y
-CONFIG_FS_PROCFS=y
-CONFIG_IDLETHREAD_STACKSIZE=4194304
-CONFIG_INIT_ENTRYPOINT="nsh_main"
-CONFIG_INTEL64_ONESHOT=y
-CONFIG_IOB_ALIGNMENT=64
-CONFIG_IOB_BUFSIZE=2048
-CONFIG_IOB_NBUFFERS=1024
-CONFIG_IOB_NCHAINS=255
-CONFIG_LIBM=y
-CONFIG_NET=y
-CONFIG_NETDB_DNSCLIENT=y
-CONFIG_NETDEV_LATEINIT=y
-CONFIG_NETDEV_WORK_THREAD=y
-CONFIG_NETINIT_DRIPADDR=0xc0a80801
-CONFIG_NETINIT_IPADDR=0x0a000102
-CONFIG_NETUTILS_CODECS=y
-CONFIG_NETUTILS_IPERF=y
-CONFIG_NETUTILS_TELNETD=y
-CONFIG_NETUTILS_TFTPC=y
-CONFIG_NETUTILS_WEBCLIENT=y
-CONFIG_NET_BROADCAST=y
-CONFIG_NET_E1000=y
-CONFIG_NET_ETH_PKTSIZE=1514
-CONFIG_NET_ICMP_SOCKET=y
-CONFIG_NET_IGC=y
-CONFIG_NET_LL_GUARDSIZE=32
-CONFIG_NET_MAX_LISTENPORTS=8
-CONFIG_NET_PKT=y
-CONFIG_NET_STATISTICS=y
-CONFIG_NET_TCP=y
-CONFIG_NET_UDP=y
-CONFIG_NSH_ARCHINIT=y
-CONFIG_NSH_BUILTIN_APPS=y
-CONFIG_NSH_READLINE=y
-CONFIG_ONESHOT=y
-CONFIG_PCI=y
-CONFIG_PCI_MSIX=y
-CONFIG_PCI_QEMU_EDU=y
-CONFIG_PCI_QEMU_TEST=y
-CONFIG_PREALLOC_CHILDSTATUS=16
-CONFIG_PRIORITY_INHERITANCE=y
-CONFIG_PTHREAD_MUTEX_TYPES=y
-CONFIG_PTHREAD_STACK_MIN=4194304
-CONFIG_RAM_SIZE=268435456
-CONFIG_SCHED_CHILD_STATUS=y
-CONFIG_SCHED_HAVE_PARENT=y
-CONFIG_SCHED_HPWORK=y
-CONFIG_SCHED_IRQMONITOR=y
-CONFIG_SCHED_LPWORK=y
-CONFIG_SCHED_TICKLESS=y
-CONFIG_SCHED_TICKLESS_ALARM=y
-CONFIG_SCHED_TICKLESS_LIMIT_MAX_SLEEP=y
-CONFIG_SCHED_WAITPID=y
-CONFIG_SIG_DEFAULT=y
-CONFIG_SMP=y
-CONFIG_START_DAY=3
-CONFIG_START_MONTH=3
-CONFIG_START_YEAR=2011
-CONFIG_SYSTEM_CLE=y
-CONFIG_SYSTEM_NSH=y
-CONFIG_SYSTEM_TIME64=y
-CONFIG_TESTING_OSTEST=y
-CONFIG_TESTING_OSTEST_STACKSIZE=4194304
-CONFIG_TESTING_SMP=y
-CONFIG_USEC_PER_TICK=1
