--- boards/x86_64/intel64/qemu-intel64/configs/lvgl/defconfig
@@ -1,73 +1,0 @@
-#
-# This file is autogenerated: PLEASE DO NOT EDIT IT.
-#
-# You can use "make menuconfig" to make any modifications to the installed .config file.
-# You can then do "make savedefconfig" to generate a new defconfig file that includes your
-# modifications.
-#
-# CONFIG_ARCH_PVHBOOT is not set
-CONFIG_16550_ADDRWIDTH=16
-CONFIG_16550_UART0=y
-CONFIG_16550_UART0_BASE=0x3f8
-CONFIG_16550_UART0_CLOCK=1843200
-CONFIG_16550_UART0_IRQ=36
-CONFIG_16550_UART0_RXBUFSIZE=16
-CONFIG_16550_UART0_SERIAL_CONSOLE=y
-CONFIG_16550_UART0_TXBUFSIZE=16
-CONFIG_16550_UART=y
-CONFIG_ARCH="x86_64"
-CONFIG_ARCH_BOARD="qemu-intel64"
-CONFIG_ARCH_BOARD_INTEL64_QEMU=y
-CONFIG_ARCH_CHIP="intel64"
-CONFIG_ARCH_CHIP_INTEL64_QEMU=y
-CONFIG_ARCH_INTEL64_CORE_FREQ_KHZ=2600000
-CONFIG_ARCH_SIZET_LONG=y
-CONFIG_ARCH_X86_64=y
-CONFIG_BOARD_LOOPSPERMSEC=999
-CONFIG_BOOT_RUNFROMEXTSRAM=y
-CONFIG_BUILTIN=y
-CONFIG_CONSOLE_SYSLOG=y
-CONFIG_DEBUG_SYMBOLS=y
-CONFIG_DEFAULT_TASK_STACKSIZE=4194304
-CONFIG_DRIVERS_VIDEO=y
-CONFIG_EXAMPLES_LVGLDEMO=y
-CONFIG_FS_PROCFS=y
-CONFIG_GRAPHICS_LVGL=y
-CONFIG_IDLETHREAD_STACKSIZE=4194304
-CONFIG_INIT_ENTRYPOINT="nsh_main"
-CONFIG_LIBM=y
-CONFIG_LV_USE_CLIB_MALLOC=y
-CONFIG_LV_USE_CLIB_SPRINTF=y
-CONFIG_LV_USE_CLIB_STRING=y
-CONFIG_LV_USE_DEMO_WIDGETS=y
-CONFIG_LV_USE_LOG=y
-CONFIG_LV_USE_NUTTX=y
-CONFIG_LV_USE_NUTTX_TOUCHSCREEN=y
-CONFIG_MULTBOOT2_FB=y
-CONFIG_NSH_ARCHINIT=y
-CONFIG_NSH_BUILTIN_APPS=y
-CONFIG_NSH_DISABLE_IFCONFIG=y
-CONFIG_NSH_DISABLE_IFUPDOWN=y
-CONFIG_PREALLOC_CHILDSTATUS=16
-CONFIG_PRIORITY_INHERITANCE=y
-CONFIG_PTHREAD_MUTEX_DEFAULT_PRIO_INHERIT=y
-CONFIG_PTHREAD_MUTEX_TYPES=y
-CONFIG_PTHREAD_STACK_MIN=4194304
-CONFIG_RAM_SIZE=268435456
-CONFIG_SCHED_CHILD_STATUS=y
-CONFIG_SCHED_HAVE_PARENT=y
-CONFIG_SCHED_THREAD_LOCAL=y
-CONFIG_SCHED_TICKLESS=y
-CONFIG_SCHED_TICKLESS_ALARM=y
-CONFIG_SCHED_TICKLESS_LIMIT_MAX_SLEEP=y
-CONFIG_SCHED_WAITPID=y
-CONFIG_SIG_DEFAULT=y
-CONFIG_START_DAY=3
-CONFIG_START_MONTH=3
-CONFIG_START_YEAR=2011
-CONFIG_SYSTEM_NSH=y
-CONFIG_SYSTEM_TIME64=y
-CONFIG_TESTING_OSTEST=y
-CONFIG_TESTING_OSTEST_STACKSIZE=4194304
-CONFIG_USEC_PER_TICK=1
-CONFIG_VIDEO_FB=y
