--- boards/x86_64/intel64/qemu-intel64/configs/nsh_pci/defconfig
@@ -1,68 +1,0 @@
-#
-# This file is autogenerated: PLEASE DO NOT EDIT IT.
-#
-# You can use "make menuconfig" to make any modifications to the installed .config file.
-# You can then do "make savedefconfig" to generate a new defconfig file that includes your
-# modifications.
-#
-# CONFIG_ARCH_PVHBOOT is not set
-CONFIG_16550_ADDRWIDTH=16
-CONFIG_16550_PCI_UART0=y
-CONFIG_16550_PCI_UART0_CLOCK=1843200
-CONFIG_16550_PCI_UART0_DEVICE=0x9100
-CONFIG_16550_PCI_UART0_SERIAL_CONSOLE=y
-CONFIG_16550_PCI_UART0_VENDOR=0x125b
-CONFIG_16550_PCI_UART=y
-CONFIG_ARCH="x86_64"
-CONFIG_ARCH_BOARD="qemu-intel64"
-CONFIG_ARCH_BOARD_INTEL64_QEMU=y
-CONFIG_ARCH_CHIP="intel64"
-CONFIG_ARCH_CHIP_INTEL64_QEMU=y
-CONFIG_ARCH_INTEL64_CORE_FREQ_KHZ=0
-CONFIG_ARCH_SIZET_LONG=y
-CONFIG_ARCH_X86_64=y
-CONFIG_BOARDCTL_RESET=y
-CONFIG_BOARD_LOOPSPERMSEC=999
-CONFIG_BOOT_RUNFROMEXTSRAM=y
-CONFIG_BUILTIN=y
-CONFIG_DEBUG_FEATURES=y
-CONFIG_DEBUG_FULLOPT=y
-CONFIG_DEBUG_SYMBOLS=y
-CONFIG_DEFAULT_TASK_STACKSIZE=4194304
-CONFIG_DEV_SIMPLE_ADDRENV=y
-CONFIG_EXAMPLES_HELLO=y
-CONFIG_EXAMPLES_ONESHOT=y
-CONFIG_EXPERIMENTAL=y
-CONFIG_FS_PROCFS=y
-CONFIG_IDLETHREAD_STACKSIZE=4194304
-CONFIG_INIT_ENTRYPOINT="nsh_main"
-CONFIG_INTEL64_HPET_MIN_DELAY=10
-CONFIG_INTEL64_ONESHOT=y
-CONFIG_LIBM=y
-CONFIG_NSH_ARCHINIT=y
-CONFIG_NSH_BUILTIN_APPS=y
-CONFIG_ONESHOT=y
-CONFIG_PCI=y
-CONFIG_PCI_MSIX=y
-CONFIG_PREALLOC_CHILDSTATUS=16
-CONFIG_PRIORITY_INHERITANCE=y
-CONFIG_PTHREAD_MUTEX_DEFAULT_PRIO_INHERIT=y
-CONFIG_PTHREAD_MUTEX_TYPES=y
-CONFIG_PTHREAD_STACK_MIN=4194304
-CONFIG_RAM_SIZE=268435456
-CONFIG_SCHED_CHILD_STATUS=y
-CONFIG_SCHED_HAVE_PARENT=y
-CONFIG_SCHED_HPWORK=y
-CONFIG_SCHED_LPWORK=y
-CONFIG_SCHED_TICKLESS=y
-CONFIG_SCHED_TICKLESS_ALARM=y
-CONFIG_SCHED_TICKLESS_LIMIT_MAX_SLEEP=y
-CONFIG_SCHED_WAITPID=y
-CONFIG_SIG_DEFAULT=y
-CONFIG_START_DAY=3
-CONFIG_START_MONTH=3
-CONFIG_START_YEAR=2011
-CONFIG_SYSLOG_DEVPATH="/dev/S1"
-CONFIG_SYSTEM_NSH=y
-CONFIG_SYSTEM_TIME64=y
-CONFIG_TESTING_OSTEST=y
