--- boards/x86_64/intel64/qemu-intel64/src/qemu_bringup.c
@@ -1,115 +1,0 @@
-/****************************************************************************
- * boards/x86_64/intel64/qemu-intel64/src/qemu_bringup.c
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.  The
- * ASF licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- ****************************************************************************/
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#include <nuttx/config.h>
-
-#include <stdbool.h>
-#include <stdio.h>
-#include <debug.h>
-#include <errno.h>
-
-#include <nuttx/board.h>
-#include <nuttx/fs/fs.h>
-
-#ifdef CONFIG_ONESHOT
-#  include <nuttx/timers/oneshot.h>
-#endif
-
-#ifdef CONFIG_PCI
-#  include <nuttx/pci/pci.h>
-#endif
-
-#ifdef CONFIG_VIDEO_FB
-#  include <nuttx/video/fb.h>
-#endif
-
-#include "qemu_intel64.h"
-
-/****************************************************************************
- * Pre-processor Definitions
- ****************************************************************************/
-
-#ifdef CONFIG_ARCH_INTEL64_HPET_ALARM
-#  if CONFIG_ARCH_INTEL64_HPET_ALARM_CHAN != 0
-#    error this logic requires that HPET_ALARM_CHAN is set to 0
-#  endif
-#  define ONESHOT_TIMER 1
-#else
-#  define ONESHOT_TIMER 0
-#endif
-
-/****************************************************************************
- * Public Functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: qemu_bringup
- ****************************************************************************/
-
-int qemu_bringup(void)
-{
-#ifdef CONFIG_ONESHOT
-  struct oneshot_lowerhalf_s *os = NULL;
-#endif
-
-  int ret = OK;
-
-#ifdef CONFIG_PCI
-  /* Register the PCI bus drivers */
-
-  pci_register_drivers();
-#endif
-
-#ifdef CONFIG_FS_PROCFS
-  /* Mount the procfs file system */
-
-  ret = nx_mount(NULL, "/proc", "procfs", 0, NULL);
-  if (ret < 0)
-    {
-      serr("ERROR: Failed to mount procfs at %s: %d\n", "/proc", ret);
-    }
-#endif
-
-#ifdef CONFIG_ONESHOT
-  os = oneshot_initialize(ONESHOT_TIMER, 10);
-  if (os)
-    {
-      oneshot_register("/dev/oneshot", os);
-    }
-#endif
-
-#ifdef CONFIG_VIDEO_FB
-  /* Initialize and register the framebuffer driver */
-
-  ret = fb_register(0, 0);
-  if (ret < 0)
-    {
-      syslog(LOG_ERR, "ERROR: fb_register() failed: %d\n", ret);
-    }
-#endif
-
-  return ret;
-}
