--- boards/x86_64/intel64/qemu-intel64/src/romfs.h
@@ -1,39 +1,0 @@
-/****************************************************************************
- * boards/x86_64/intel64/qemu-intel64/src/romfs.h
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.  The
- * ASF licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- ****************************************************************************/
-
-#ifndef __BOARDS_X86_64_INTEL64_QEMU_INTEL64_SRC_ROMFS_H
-#define __BOARDS_X86_64_INTEL64_QEMU_INTEL64_SRC_ROMFS_H
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-/****************************************************************************
- * Pre-processor Definitions
- ****************************************************************************/
-
-/****************************************************************************
- * Public Function Prototypes
- ****************************************************************************/
-
-extern const unsigned char romfs_img[];
-extern const unsigned int  romfs_img_len;
-
-#endif /* __BOARDS_X86_64_INTEL64_QEMU_INTEL64_SRC_ROMFS_H */
