--- boards/x86_64/intel64/qemu-intel64/src/romfs_stub.c
@@ -1,39 +1,0 @@
-/****************************************************************************
- * boards/x86_64/intel64/qemu-intel64/src/romfs_stub.c
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.  The
- * ASF licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- ****************************************************************************/
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#include <nuttx/compiler.h>
-
-/****************************************************************************
- * Public Data
- ****************************************************************************/
-
-weak_data const unsigned char aligned_data(4) romfs_img[] =
-{
-  0x00
-};
-weak_data const unsigned int romfs_img_len = 1;
-
-/****************************************************************************
- * Public Functions
- ****************************************************************************/
