+++ boards/x86_64/qemu/qemu-intel64/configs/knsh_romfs_pci/defconfig
@@ -1,0 +1,98 @@
+#
+# This file is autogenerated: PLEASE DO NOT EDIT IT.
+#
+# You can use "make menuconfig" to make any modifications to the installed .config file.
+# You can then do "make savedefconfig" to generate a new defconfig file that includes your
+# modifications.
+#
+# CONFIG_ARCH_X86_64_ACPI_DUMP is not set
+# CONFIG_DEV_RANDOM is not set
+# CONFIG_SIG_SIGKILL_ACTION is not set
+# CONFIG_SIG_SIGSTOP_ACTION is not set
+# CONFIG_TLS_ALIGNED is not set
+# CONFIG_ARCH_PVHBOOT is not set
+CONFIG_16550_ADDRWIDTH=16
+CONFIG_16550_PCI_UART0=y
+CONFIG_16550_PCI_UART0_CLOCK=1843200
+CONFIG_16550_PCI_UART0_DEVICE=0x9100
+CONFIG_16550_PCI_UART0_SERIAL_CONSOLE=y
+CONFIG_16550_PCI_UART0_VENDOR=0x125b
+CONFIG_16550_PCI_UART=y
+CONFIG_16550_PCI_UART_AX99100=y
+CONFIG_ALOG=-1
+CONFIG_ARCH="x86_64"
+CONFIG_ARCH_ADDRENV=y
+CONFIG_ARCH_BOARD="qemu-intel64"
+CONFIG_ARCH_BOARD_INTEL64_QEMU=y
+CONFIG_ARCH_CHIP="qemu"
+CONFIG_ARCH_CHIP_INTEL64_QEMU=y
+CONFIG_ARCH_DATA_VBASE=0x800080000
+CONFIG_ARCH_HEAP_VBASE=0x8000f0000
+CONFIG_ARCH_INTEL64_CORE_FREQ_KHZ=0
+CONFIG_ARCH_KERNEL_STACKSIZE=8192
+CONFIG_ARCH_PGPOOL_MAPPING=y
+CONFIG_ARCH_PGPOOL_PBASE=0x00c000000
+CONFIG_ARCH_PGPOOL_SIZE=67108864
+CONFIG_ARCH_PGPOOL_VBASE=0x10c000000
+CONFIG_ARCH_SIZET_LONG=y
+CONFIG_ARCH_TEXT_VBASE=0x800000000
+CONFIG_ARCH_X86_64=y
+CONFIG_BOARDCTL_APP_SYMTAB=y
+CONFIG_BOARDCTL_ROMDISK=y
+CONFIG_BOARD_LATE_INITIALIZE=y
+CONFIG_BOARDCTL_RESET=y
+CONFIG_BOARD_LOOPSPERMSEC=999
+CONFIG_BOOT_RUNFROMEXTSRAM=y
+CONFIG_BUILD_KERNEL=y
+CONFIG_DEBUG_ASSERTIONS=y
+CONFIG_DEBUG_ASSERTIONS_EXPRESSION=y
+CONFIG_DEBUG_FEATURES=y
+CONFIG_DEBUG_SYMBOLS=y
+CONFIG_DEFAULT_TASK_STACKSIZE=16384
+CONFIG_DEV_SIMPLE_ADDRENV=y
+CONFIG_ELF=y
+CONFIG_ELF_STACKSIZE=163840
+CONFIG_EXAMPLES_HELLO=m
+CONFIG_EXPERIMENTAL=y
+CONFIG_FS_PROCFS=y
+CONFIG_FS_ROMFS=y
+CONFIG_IDLETHREAD_STACKSIZE=16384
+CONFIG_INIT_FILEPATH="/system/bin/init"
+CONFIG_INIT_MOUNT=y
+CONFIG_INIT_MOUNT_FLAGS=0x1
+CONFIG_INIT_MOUNT_TARGET="/system/bin"
+CONFIG_LIBC_EXECFUNCS=y
+CONFIG_INTEL64_HPET_MIN_DELAY=10
+CONFIG_LIBM=y
+CONFIG_MM_KERNEL_HEAPSIZE=16384
+CONFIG_MM_PGALLOC=y
+CONFIG_NSH_ARCHINIT=y
+CONFIG_PCI=y
+CONFIG_PCI_MSIX=y
+CONFIG_NSH_DISABLE_IFCONFIG=y
+CONFIG_NSH_DISABLE_IFUPDOWN=y
+CONFIG_NSH_FILE_APPS=y
+CONFIG_PREALLOC_CHILDSTATUS=16
+CONFIG_PTHREAD_MUTEX_TYPES=y
+CONFIG_PTHREAD_STACK_MIN=16384
+CONFIG_RAM_SIZE=268435456
+CONFIG_RAM_VSTART=0x100000000
+CONFIG_SCHED_CHILD_STATUS=y
+CONFIG_SCHED_HAVE_PARENT=y
+CONFIG_SCHED_HPWORK=y
+CONFIG_SCHED_LPWORK=y
+CONFIG_SCHED_TICKLESS=y
+CONFIG_SCHED_TICKLESS_ALARM=y
+CONFIG_SCHED_TICKLESS_LIMIT_MAX_SLEEP=y
+CONFIG_SCHED_WAITPID=y
+CONFIG_SIG_DEFAULT=y
+CONFIG_START_DAY=3
+CONFIG_START_MONTH=3
+CONFIG_START_YEAR=2011
+CONFIG_SYSLOG_DEVPATH="/dev/S1"
+CONFIG_SYSTEM_NSH=y
+CONFIG_SYSTEM_NSH_PROGNAME="init"
+CONFIG_SYSTEM_TIME64=y
+CONFIG_TESTING_OSTEST=y
+CONFIG_TESTING_OSTEST_FPUTESTDISABLE=y
+CONFIG_TESTING_OSTEST_STACKSIZE=16384
