+++ boards/x86_64/qemu/qemu-intel64/src/Makefile
@@ -1,0 +1,36 @@
+############################################################################
+# boards/x86_64/qemu/qemu-intel64/src/Makefile
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.  The
+# ASF licenses this file to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance with the
+# License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+############################################################################
+
+include $(TOPDIR)/Make.defs
+
+CSRCS = qemu_boot.c qemu_bringup.c qemu_net.c
+CSRCS += $(if $(wildcard romfs_boot.c), romfs_boot.c, romfs_stub.c)
+
+ifeq ($(CONFIG_BOARDCTL),y)
+  CSRCS += qemu_appinit.c
+endif
+
+ifeq ($(CONFIG_BOARDCTL_RESET),y)
+  CSRCS += qemu_reset.c
+endif
+
+include $(TOPDIR)/boards/Board.mk
